src/server/pyproject.toml
[tool.poetry]
name = "cognition-hackathon"
version = "0.1.0"
description = "ML Experiment Runner with Modal Serverless Infrastructure"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
jupyter = "^1.0.0"
notebook = "^7.0.0"
ipykernel = "^6.25.0"
modal = "^1.1.0"
torch = ">=2.0.0"
torchvision = ">=0.15.0"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.0"
pandas = "^2.1.0"
numpy = "^1.24.0"
tqdm = "^4.66.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
httpx = "^0.25.0"
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
anthropic = "^0.62.0"
python-dotenv = "^1.0.0"
redis = "^5.0.0"
openai = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.9.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"
ipywidgets = "^8.1.0"
nbconvert = "^7.9.0"
jupyter = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true